--- SimuladorExamenUPN.Tests/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.12.0" targetFramework="net452" />
-  <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net452" />
-  <package id="Selenium.WebDriver.ChromeDriver" version="77.0.3865.4000" targetFramework="net452" />
-</packages>
