+++ SimuladorExamenUPN.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimuladorExamenUPN", "SimuladorExamenUPN\SimuladorExamenUPN.csproj", "{2024BD2B-5897-4C36-AB3A-738F23148507}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimuladorExamenUPN.Tests", "SimuladorExamenUPN.Tests\SimuladorExamenUPN.Tests.csproj", "{D592F2A4-6B8A-4376-9023-5F46AF78645E}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{2024BD2B-5897-4C36-AB3A-738F23148507}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2024BD2B-5897-4C36-AB3A-738F23148507}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2024BD2B-5897-4C36-AB3A-738F23148507}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D592F2A4-6B8A-4376-9023-5F46AF78645E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D592F2A4-6B8A-4376-9023-5F46AF78645E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D592F2A4-6B8A-4376-9023-5F46AF78645E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D592F2A4-6B8A-4376-9023-5F46AF78645E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
