+++ SimuladorExamenUPN/Controllers/HomeController.cs
@@ -1,64 +1,51 @@
-﻿using System;
+﻿using SimuladorExamenUPN.DB;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using System.Data.Entity;
 
 namespace SimuladorExamenUPN.Controllers
 {
     [Authorize]
     public class HomeController : Controller
     {
-        [HttpGet]
+        // GET: HomeExamen
+        SimuladorContext db;
+        public HomeController()
+        {
+            db = new SimuladorContext();
+        }
         public ActionResult Index()
         {
-            if (Session["usuario"] == null)
-                return RedirectToAction("Login", "usuario");
-
-            return View();
+            var examenes = db.Examenes
+                .Include(o => o.Tema.Categorias.Select(s =>s.Categoria))
+                .Include(o =>o.Usuario)
+                .Where(o => o.EstaActivo == true)
+                .ToList();
+            return View(examenes);
         }
 
-        public ActionResult About()
+        public ActionResult Confirmar(int ExamenId)
         {
-            ViewBag.Message = "Your application description page.";
-
-            return View();
+            var examen = db.Examenes
+                .Where(o => o.Id == ExamenId)
+                .Include(o =>o.Tema)
+                .Include(u => u.Usuario)
+                .FirstOrDefault();
+            return View(examen);
         }
 
-        public ActionResult Contact()
+        public ActionResult DarExamen(int ExamenId)
         {
-            ViewBag.Message = "Your contact page.";
+            
+            var examen = db.Examenes.Where(o => o.Id == ExamenId)
+                .Include(o => o.Preguntas.Select(s => s.Pregunta.Alternativas))
+                .FirstOrDefault();
 
-            return View();
+            return View(examen);
         }
-
-        public ViewResult Page1() {
-            return View();
-        }
-
-        public ViewResult Page2()
-        {
-            return View();
-        }
-
-        public ViewResult Ciudades(int pais)
-        {
-            var ciudades = GetCiudades(pais);
-            return View(ciudades);
-        }
-
-        private List<String> GetCiudades(int pais)
-        {
-            //return db.Ciudades.Where(o => o.PaisId == pais).ToList()
-            switch(pais)
-            {
-                case 1:
-                    return new List<string> { "Lima", "Cajamarca", "La Libertad" };
-                case 2:
-                    return new List<string> { "Bueno Aires", "Mendoza", "Rosario" };
-                default:
-                    return new List<string>();
-            }
-        }
+        
     }
 }
