--- SimuladorExamenUPN/Views/Home/Ciudades.cshtml
@@ -1,12 +1,0 @@
-﻿
-@{
-    Layout = null;
-}
-
-<select>
-    @foreach (var item in Model)
-    {
-        <option>@item</option>
-    }
-</select>
-
