+++ SimuladorExamenUPN/Views/Home/Confirmar.cshtml
@@ -8,5 +8,5 @@
 <h3>¿Seguro desea iniciar el examen de @Model.Tema.Nombre ? </h3>
 <h3>Creado por @Model.Usuario.Nombres</h3>
 
-<a class="btn-danger" href="/HomeExamen/DarExamen?ExamenId=@Model.Id">Iniciar</a>
-<a class="btn-danger" href="#">Cancelar</a>
+<a class="btn btn-danger" href="/Home/DarExamen?ExamenId=@Model.Id">Iniciar</a>
+<a class="btn btn-link" href="#">Cancelar</a>
