+++ SimuladorExamenUPN/Views/Home/DarExamen.cshtml
@@ -7,7 +7,7 @@
 
 <h2>DarExamen</h2>
 
-<form action="/HomeExamen/GuardarExamen" method="post">
+<form action="/Home/GuardarExamen" method="post">
     <input type="hidden" name="ExamenId"value="@Model.Id" />
     <table class="table table-borderless">
         <thead>
