--- SimuladorExamenUPN/Views/Home/Page1.cshtml
@@ -1,30 +1,0 @@
-﻿@{Layout = null;}
-
-<script src="~/Scripts/jquery-1.10.2.min.js"></script>
-
-
-<select id="pais">
-    <option value="">--- Seleccione Pais --</option>
-    <option value="1">Perú</option>
-    <option value="2">Argentina</option>
-</select>
-
-<div id="selectCiudades">
-
-</div>
-
-
-<script>
-  
-    $("#pais").change(function () {
-        var valor = $("#pais").val();
-
-        $.ajax({ url: "/home/ciudades?pais=" + valor })
-        .done(function (res) {
-            $("#selectCiudades").html(res);
-        });
-
-    });
-
-
-</script>
