--- SimuladorExamenUPN/Views/Home/Page2.cshtml
@@ -1,2 +1,0 @@
-﻿@{Layout = null;}
-<p>lorem ipsum</p>
