+++ SimuladorExamenUPN/Views/Pregunta/Crear.cshtml
@@ -1,4 +1,4 @@
-﻿
+﻿@model SimuladorExamenUPN.Models.Pregunta
 
 <h2>@ViewBag.Tema.Nombre</h2>
 <h4>Crear Pregunta</h4>
@@ -8,21 +8,20 @@
     <div class="form-group">
         <label>Pregunta</label>
         <textarea class="form-control" name="Descripcion">@Model.Descripcion</textarea>
-        <input type="checkbox" name="Hello" value="World" />
         @Html.ValidationMessage("Descripcion")
     </div>
     <div class="form-group">
         <label>Alternativas</label>
         @*<div class="row">
-            <div class="col-11">
-                <input class="form-control" id="inputAlternative" type="text" value="" />
-            </div>
-            <div class="col-1">
-                <a href="#" id="btnAddAlternativa" class="btn btn-success">Añadir</a>
-            </div>
-        </div>*@
-        <br />     
-        
+                <div class="col-11">
+                    <input class="form-control" id="inputAlternative" type="text" value="" />
+                </div>
+                <div class="col-1">
+                    <a href="#" id="btnAddAlternativa" class="btn btn-success">Añadir</a>
+                </div>
+            </div>*@
+        <br />
+
         @Html.ValidationMessage("Alternativas")
         <table class="table table-condensed" id="tableAlternativas">
             @{ var index = 0; }
@@ -33,17 +32,21 @@
                         <input type="text" class="form-contro" name="Alternativas[@index].Descripcion" value="@item.Descripcion">
                     </td>
                     <td>
-                        <input type="checkbox" name="Alternativas[@index].EsCorrecto" value="true" @(item.EsCorrecto ? "checked": "") />
+                        <input type="checkbox" name="Alternativas[@index].EsCorrecto" value="true" @(item.EsCorrecto ? "checked" : "") />
                         <input type="hidden" name="Alternativas[@index].EsCorrecto" value="false"> ¿Es Correcta?
                     </td>
                 </tr>
                 index++;
             }
 
-         
-            
         </table>
-        <a href="#" id="btnAddAlternativa" class="btn btn-success">Añadir</a>
+        @if (!Model.Alternativas.Any())
+        {
+            <div class="alert alert-primary" id="AlertEmpty">
+                No se han añadido alternativas aun
+            </div>
+        }
+        <a href="#" id="btnAddAlternativa" class="btn btn-success">Añadir anternativa</a>
     </div>
 
     <button class="btn btn-primary">Guardar</button>
@@ -51,6 +54,7 @@
 </form>
 
 <script>
+    var $alert = $("#AlertEmpty");
     var index = $("#tableAlternativas").find("tr").length;
 
     $("#btnAddAlternativa").click(function (event) {
@@ -72,6 +76,7 @@
         
         inputAlternativa.val("");
         index++;
+        $alert.hide();
 
         
     });
