+++ SimuladorExamenUPN/Views/Shared/_Layout.cshtml
@@ -20,10 +20,7 @@
         </button>
 
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <ul class="navbar-nav mr-auto">
-                <li class="nav-item">
-                    <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-                </li>
+            <ul class="navbar-nav mr-auto">               
                 <li class="nav-item">
                     <a class="nav-link" id="temaLink" href="/Tema">Temas</a>
                 </li>
