+++ SimuladorExamenUPN/Views/Usuario/Login.cshtml
@@ -15,12 +15,12 @@
 
             <div class="form-group">
                 <label>Usuario</label>
-                <input class="form-control" type="text" name="Username" />
+                <input class="form-control" type="text" name="Username" value="admin" />
             </div>
 
             <div class="form-group">
                 <label>Password</label>
-                <input class="form-control" type="password" name="Password" />
+                <input class="form-control" type="password" name="Password" value="admin" />
             </div>
 
             <button class="btn btn-primary">Iniciar Sesión</button>
